[gh-9] str_to_utf8 method removed
It is impossible to 100% accurately guess the encoding of a string and convert it based on that guess. This function is unnecessary in properly designed applications which specify character encoding for input and output.

If the encoding is known and a conversion required, `iconv` or `mb_convert_encoding` are the right tools for the job.